Add onErrorRetry parameter
The default action is to silent the error and to break the sync cycle,
so backward compatibility is kept.

If 'onErrorRetry' is set to true, the sync cycle continues whatever
happens during the http call or the during the execution of the
watchers.

'onErrorRetry' could also be a function, allowing to call 'nextSync()'
on the synchronization object to delay a new run of synchronization, by
default after the interval defined in the parameters. 'nextSync()'
possibly takes an interval, which allows to override the default (for
instance we may want to extend the interval when some errors happen).